+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - `CHANGELOG.md` in the spirit of [Keep a CHANGELOG](http://keepachangelog.com/).
 - Rough `TODO.md` listing some things that should be done.
 - Error dialog if Lilith is started with Java prior to 1.8.0_45.
+- Added "Clear view" shortcut Cmd-K. K like clear.
 
 ### Changed
 - Requires Java 8.